Add database auto-replication using edge-server-tools
