+++ package.json
@@ -53,6 +53,7 @@
     "cleaners": "^0.2.1",
     "cookie-parser": "^1.4.5",
     "cors": "^2.8.5",
+    "edge-server-tools": "^0.1.0",
     "express": "^4.17.1",
     "nano": "^8.2.2",
     "node-fetch": "^2.6.1",
