+++ src/couchSchema.ts
@@ -1,6 +1,29 @@
 /* eslint-disable no-var */
 
-import { CouchDbInfo } from './util/rebuildCouch'
+/**
+ * Describes a single database that should exist.
+ */
+export interface CouchDbInfo {
+  name: string
+  indexes?: Array<{
+    index: {
+      fields: string[]
+      partial_filter_selector?: any
+    }
+    ddoc: string
+    name: string
+    type: 'json'
+  }>
+  views?: Array<{
+    name: string
+    views: {
+      [viewName: string]: {
+        map?: string
+        reduce?: string
+      }
+    }
+  }>
+}
 
 export const couchSchema: CouchDbInfo[] = [
   {
