+++ src/indexApi.ts
@@ -9,14 +9,19 @@ import {
 import cluster from 'cluster'
 import cookieParser from 'cookie-parser'
 import cors from 'cors'
+import {
+  autoReplication,
+  forkChildren,
+  makePeriodicTask,
+  rebuildCouch
+} from 'edge-server-tools'
 import express from 'express'
 import nano from 'nano'
-import { cpus } from 'os'
 
 import config from '../config.json'
 import { couchSchema } from './couchSchema'
-import { rebuildCouch } from './util/rebuildCouch'
 
+const AUTOREPLICATION_DELAY = 1000 * 60 * 30 // 30 minutes
 const FIVE_MINUTES = 1000 * 60 * 5
 
 const asLog = asObject({
@@ -116,7 +121,7 @@ interface Selector {
 
 const nanoDb = nano(config.couchDbFullpath)
 
-function main(): void {
+function api(): void {
   // start express and couch db server
   const app = express()
   const logsRecords = nanoDb.use('logs_records')
@@ -286,34 +291,25 @@ function main(): void {
   })
 }
 
-const numCPUs = cpus().length
-
-if (cluster.isMaster) {
-  rebuildCouch(config.couchDbFullpath, couchSchema)
-    .then(() => {
-      const instanceCount =
-        config.instanceCount != null ? config.instanceCount ?? numCPUs : numCPUs
-
-      // Fork workers.
-      for (let i = 0; i < instanceCount; i++) {
-        cluster.fork()
-      }
-
-      // Restart workers when they exit
-      cluster.on('exit', (worker, code, signal) => {
-        console.log(
-          `Worker ${worker.process.pid} died with code ${code} and signal ${signal}`
-        )
-        console.log(`Forking new worker process...`)
-        cluster.fork()
-      })
-    })
-    .catch(failStartup)
-} else {
-  main()
+async function main(): Promise<void> {
+  const { couchDbFullpath, infoServerAddress, infoServerApiKey } = config
+  if (cluster.isMaster) {
+    await rebuildCouch(couchDbFullpath, couchSchema).catch(e => console.log(e))
+    const task = makePeriodicTask(
+      async () =>
+        await autoReplication(
+          infoServerAddress,
+          'logsServer',
+          infoServerApiKey,
+          couchDbFullpath
+        ).catch(e => console.log(e)),
+      AUTOREPLICATION_DELAY
+    )
+    task.start()
+    forkChildren()
+  } else {
+    api()
+  }
 }
 
-function failStartup(err: any): void {
-  console.error(err)
-  process.exit(1)
-}
+main().catch(e => console.log(e))
